requirements.txt
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Cors==3.0.10
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
PyJWT==2.6.0
PyMySQL==1.0.2
pymysql-pool==0.3.7
scikit-learn==1.2.2
